Removed tax rates import from datagrid
