+++ packages/Webkul/Admin/src/Resources/lang/pt_BR/app.php
@@ -2513,16 +2513,6 @@ return [
                     'tax-rate'     => 'Taxa de Imposto',
                     'title'        => 'Taxas de Imposto',
 
-                    'import' => [
-                        'duplicate-error'  => 'O identificador deve ser único, identificador duplicado :identifier na linha :position.',
-                        'enough-row-error' => 'O arquivo não tem linhas suficientes',
-                        'import-btn'       => 'Importar',
-                        'title'            => 'Upload',
-                        'upload-error'     => 'O arquivo deve ser do tipo: xls, xlsx, csv.',
-                        'upload-success'   => 'Taxa de imposto carregada com sucesso',
-                        'validation'       => 'Tipo permitido: xls, xlsx, csv.',
-                    ],
-
                     'datagrid' => [
                         'country'    => 'País',
                         'delete'     => 'Excluir',
