+++ packages/Webkul/Admin/src/Resources/lang/sin/app.php
@@ -2513,16 +2513,6 @@ return [
                     'tax-rate'     => 'බදා අනුපාත',
                     'title'        => 'බදා අනුපාත',
 
-                    'import' => [
-                        'duplicate-error'  => 'හැඳුනුම්කරු විවෘත විය යුතුය, පේළිය:position හි අනුපාත හැඳුනුම්කරු :identifier.',
-                        'enough-row-error' => 'ගොනුවේ සුලු පේළියෙන් පසු සහිත විස්තර නොමැත',
-                        'import-btn'       => 'ආයුක්ත',
-                        'title'            => 'උඩුගත',
-                        'upload-error'     => 'ගොනුවේ වර්ගය: xls, xlsx, csv යෙදී හෝදින් හැකි යැයි විය හැක.',
-                        'upload-success'   => 'බදු අනුවාදය සාර්ථකව උඩුගත වී ඇත',
-                        'validation'       => 'ඉඩ දෙන වර්ගය: xls, xlsx, csv.',
-                    ],
-
                     'datagrid' => [
                         'country'    => 'රට',
                         'delete'     => 'මකන්න',
