+++ packages/Webkul/Admin/src/Resources/views/settings/taxes/rates/index.blade.php
@@ -10,13 +10,6 @@
         </p>
 
         <div class="flex gap-x-2.5 items-center">
-            <!-- Tax Rate Emport -->
-            <v-tax-rates-import>
-                <button class="transparent-button hover:bg-gray-200 dark:hover:bg-gray-800 dark:text-white">
-                    @lang('admin::app.settings.taxes.rates.index.import.import-btn')
-                </button>
-            </v-tax-rates-import>
-
             <!-- Tax Rate Export -->
             <x-admin::datagrid.export src="{{ route('admin.settings.taxes.rates.index') }}" />
 
@@ -33,106 +26,4 @@
         :src="route('admin.settings.taxes.rates.index')" 
         ref="datagrid"
     />
-
-    @pushOnce('scripts')
-        <script
-            type="text/x-template"
-            id="v-tax-rates-import-template"
-        >
-            <div>
-                <!-- Tax Rate Import Button -->
-                <button
-                    href="{{ route('admin.settings.taxes.rates.import') }}"
-                    class="transparent-button hover:bg-gray-200 dark:hover:bg-gray-800 dark:text-white"
-                    @click="$refs.importTaxRates.toggle()"
-                >
-                    @lang('admin::app.settings.taxes.rates.index.import.import-btn')
-                </button>
-
-                <!-- Modal Form -->
-                <x-admin::form
-                    v-slot="{ meta, errors, handleSubmit }"
-                    as="div"
-                >
-                    <form @submit="handleSubmit($event, create)">
-                        <x-admin::modal ref="importTaxRates">
-                            <!-- Modal Header -->
-                            <x-slot:header>
-                                <p class="text-lg text-gray-800 dark:text-white font-bold">
-                                    @lang('admin::app.settings.taxes.rates.index.import.title')
-                                </p>
-                            </x-slot>
-
-                            <!-- Modal Content -->
-                            <x-slot:content>
-                                <x-admin::form.control-group>
-                                    <x-admin::form.control-group.control
-                                        type="file"
-                                        id="file"
-                                        name="file"
-                                        ref="importFile"
-                                    />
-
-                                    <x-admin::form.control-group.error control-name="file" />
-                                </x-admin::form.control-group>
-
-                                <p class="text-xs text-gray-600 dark:text-gray-300">  
-                                    @lang('admin::app.settings.taxes.rates.index.import.validation') 
-                                </p>
-                            </x-slot>
-
-                            <!-- Modal Footer -->
-                            <x-slot:footer>
-                                <div class="flex gap-x-2.5 items-center">
-                                    <button
-                                        type="submit"
-                                        class="primary-button"
-                                    >
-                                        @lang('admin::app.settings.taxes.rates.index.import.import-btn')
-                                    </button>
-                                </div>
-                            </x-slot>
-                        </x-admin::modal>
-                    </form>
-                </x-admin::form>
-            </div>
-        </script>
-
-        <script type="module">
-            app.component('v-tax-rates-import', {
-                template: '#v-tax-rates-import-template',
-
-                methods: {
-                    create(params, {
-                        resetForm,
-                        setErrors
-                    }) {
-                        const formData = new FormData();
-
-                        const fileInput = this.$refs.importFile;
-
-                        if (fileInput.files.length) {
-                            formData.append('file', fileInput.files[0]);
-                        }
-
-                        this.$axios.post("{{ route('admin.settings.taxes.rates.import') }}", formData)
-                            .then((response) => {
-                                this.$refs.importTaxRates.toggle();
-
-                                this.$emitter.emit('add-flash', { type: 'success', message: response.data.message });
-
-                                this.$parent.$refs.datagrid.get();
-
-                                resetForm();
-                            })
-                            .catch(error => {
-                                setErrors({
-                                    file: error.response.data.message
-                                })
-                            });
-                    },
-                },
-            });
-        </script>
-    @endPushOnce
 </x-admin::layouts>
