+++ packages/Webkul/Admin/src/Routes/settings-routes.php
@@ -132,8 +132,6 @@ Route::group(['middleware' => ['admin'], 'prefix' => config('app.admin_url')], f
                 Route::put('edit/{id}', 'update')->name('admin.settings.taxes.rates.update');
 
                 Route::delete('edit/{id}', 'destroy')->name('admin.settings.taxes.rates.delete');
-
-                Route::post('import', 'import')->name('admin.settings.taxes.rates.import');
             });
         });
