+++ packages/Webkul/DataTransfer/src/Helpers/Sources/CSV.php
@@ -3,39 +3,13 @@
 namespace Webkul\DataTransfer\Helpers\Sources;
 
 use Illuminate\Support\Facades\Storage;
-use Webkul\DataTransfer\Helpers\Importers\AbstractImporter;
 
-class CSV
+class CSV extends AbstractSource
 {
     /**
      * CSV reader
      */
     protected mixed $reader;
-
-    /**
-     * Column names
-     */
-    protected array $columnNames = [];
-
-    /**
-     * Quantity of columns
-     */
-    protected int $totalColumns = 0;
-
-    /**
-     * Current row
-     */
-    protected array $currentRowData = [];
-
-    /**
-     * Current row number
-     */
-    protected int $currentRowNumber = -1;
-
-    /**
-     * Flag to indicate that wrong quote was found
-     */
-    protected bool $foundWrongQuoteFlag = false;
 
     /**
      * Create a new helper instance.
@@ -58,59 +32,7 @@ class CSV
     }
 
     /**
-     * Return the key of the current element (\Iterator interface)
-     */
-    public function getCurrentRowNumber(): int
-    {
-        return $this->currentRowNumber;
-    }
-
-    /**
-     * Checks if current position is valid
-     */
-    public function valid(): bool
-    {
-        return $this->currentRowNumber !== -1;
-    }
-
-    /**
-     * Read next line from CSV-file
-     */
-    public function current(): array
-    {
-        $row = $this->currentRowData;
-
-        if (count($row) != $this->totalColumns) {
-            if ($this->foundWrongQuoteFlag) {
-                throw new \InvalidArgumentException(AbstractImporter::ERROR_CODE_WRONG_QUOTES);
-            } else {
-                throw new \InvalidArgumentException(AbstractImporter::ERROR_CODE_COLUMNS_NUMBER);
-            }
-        }
-
-        return array_combine($this->columnNames, $row);
-    }
-
-    /**
-     * Read next line from CSV-file
-     */
-    public function next(): void
-    {
-        $this->currentRowNumber++;
-
-        $row = $this->getNextRow();
-
-        if ($row === false || $row === []) {
-            $this->currentRowData = [];
-
-            $this->currentRowNumber = -1;
-        } else {
-            $this->currentRowData = $row;
-        }
-    }
-
-    /**
-     * Read next line from CSV-file
+     * Read next line from csv
      */
     protected function getNextRow(): array
     {
@@ -132,34 +54,12 @@ class CSV
     }
 
     /**
-     * Rewind the \Iterator to the first element (\Iterator interface)
+     * Rewind the iterator to the first row
      */
     public function rewind(): void
     {
         rewind($this->reader);
 
-        $this->currentRowNumber = 0;
-
-        $this->currentRowData = [];
-
-        $this->getNextRow();
-
-        $this->next();
-    }
-
-    /**
-     * Column names getter.
-     */
-    public function getColumnNames(): array
-    {
-        return $this->columnNames;
-    }
-
-    /**
-     * Column names getter.
-     */
-    public function getTotalColumns(): int
-    {
-        return $this->columnNames;
+        parent::rewind();
     }
 }
